+++ src/Classes/Vector2D.h
@@ -30,7 +30,7 @@ public:
 	Vector2D operator*(Vector2D const& v2) { return Vector2D(x * v2.x, y * v2.y); }
 
 	Vector2D operator*(int const& i) { return Vector2D(x * i, y * i); }
-	Vector2D operator*(float const& f) { return Vector2D(x * f, y * f); }
+	Vector2D operator*(float const& f) { return Vector2D((int16_t)(x * f), (int16_t)(y * f)); }
 };
 
 struct Dir {
